merginf sending action events

+++ engine-core/World.cpp
@@ -104,7 +104,7 @@ void World::broadcastUpdates(CommsProcessor *comms) {
 
 			buffer->allocate();
 
-			((struct EventHeader*)buffer->getPointer())->type = EventType::OBJECT_UPDATE;
+			((struct EventHeader*)buffer->getPointer())->type = static_cast<int>(EventType::OBJECT_UPDATE);
 			buffer->pop();
 
 			struct ObjectUpdateHeader *ouHeader = (struct ObjectUpdateHeader*)buffer->getPointer();
